Refactor sql adapter unit tests
